Insert into the footer of your website script.js
If you want to apply this script to the entire website, insert it in the footer or footer of the page you want to apply it to. Keep in mind that you should put it in the footer only so that there is no impact on the speed of your website. When you put it on your head, you will notice some difference in speed.